Updates to JDBC JAR handling.
register-jdbc.sh now actively determines which JDBC JARs may be in-conflict with ones supplied by the user at the `/jdbc` path.  It looks up matches with existing JARs at `user-lib/jdbc` and if an in-bound linked JDBC JAR classpath matches those, this script will now remove those lingering JARs to prevent them from loading.  This can help in situations where the user might need to pin to a specific version--it ensures there is only a single JAR for a given classpath target.